+++ src/managers/ProductsManager.java
@@ -18,7 +18,7 @@ public class ProductsManager {
         product.setTitle(scanner.nextLine());
         
         System.out.println("Input product price per item: ");
-        product.setPrice(scanner.nextDouble()); //scanner.nextLine();
+        product.setPrice(scanner.nextInt()); //scanner.nextLine();
         
         System.out.println("Input products quantity: ");
         product.setQuantity(scanner.nextInt()); scanner.nextLine();
@@ -27,7 +27,7 @@ public class ProductsManager {
     
     public void productsLis(Product[] products){
         for (int i = 0; i < products.length; i++) {
-            System.out.printf("%d. Title: %s; Price: %.2f; Quantity: %d%n", i+1, products[i].getTitle(), products[i].getPrice(), products[i].getQuantity());
+            System.out.printf("%d. Title: %s; Price: %d; Quantity: %d%n", i+1, products[i].getTitle(), products[i].getPrice(), products[i].getQuantity());
         }
     }
     
@@ -43,12 +43,12 @@ public class ProductsManager {
                     products[prodNrToEdit].setTitle(scanner.nextLine());
                 break;
             }
-            System.out.printf("Edit price '%.2f'? Choose option: (y / n) ", products[prodNrToEdit].getPrice());
+            System.out.printf("Edit price '%d'? Choose option: (y / n) ", products[prodNrToEdit].getPrice());
             task = scanner.nextLine();
             switch (task.toLowerCase()){
                 case "y":
                     System.out.print("Input new price: ");
-                    products[prodNrToEdit].setPrice(scanner.nextDouble()); scanner.nextLine();
+                    products[prodNrToEdit].setPrice(scanner.nextInt()); scanner.nextLine();
                 break;
             }
             System.out.printf("Edit quantity '%d'? Choose option: (y / n) ", products[prodNrToEdit].getQuantity());
